// src/modules/XuatVien/model/xuatVienModel.ts
export interface XuatVien {
  maXuatVien: number;
  maBenhAn: number;
  maBenhNhan: string;
  hoTen: string;
  ngaySinh: string; // Hoặc có thể dùng kiểu Date nếu bạn xử lý chuyển đổi
  gioiTinh: string;
  soCMND: string;
  tinhTrang: string;
  ghiChu: string;
  trangThai: string;
  ngayRaVien: string; // Hoặc có thể dùng kiểu Date
}